# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/curd/cell/add": {
            "post": {
                "description": "新增",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cell管理"
                ],
                "summary": "新增",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Add Cell",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/curd.CellInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/curd/cell/all": {
            "get": {
                "description": "查询所有",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cell管理"
                ],
                "summary": "查询所有",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.GetAllCellResponse"
                        }
                    }
                }
            }
        },
        "/api/curd/cell/copy": {
            "post": {
                "description": "复制",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cell管理"
                ],
                "summary": "复制",
                "parameters": [
                    {
                        "description": "Copy Cell",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/curd.DelRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/curd/cell/delete": {
            "delete": {
                "description": "删除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cell管理"
                ],
                "summary": "删除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Delete Cell",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/curd.DelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/curd/cell/detail": {
            "get": {
                "description": "查询明细",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cell管理"
                ],
                "summary": "查询明细",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.GetCellDetailResponse"
                        }
                    }
                }
            }
        },
        "/api/curd/cell/enable": {
            "post": {
                "description": "禁用/启用",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cell管理"
                ],
                "summary": "禁用/启用",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Enable/Disable Cell",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/curd.EnableRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/curd/cell/export": {
            "get": {
                "description": "导出",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Cell管理"
                ],
                "summary": "导出",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "从1开始",
                        "name": "pageIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "默认每页10条",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序字段",
                        "name": "orderField",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否倒序排序",
                        "name": "desc",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Must Source",
                        "name": "mustSource",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "系统",
                        "name": "system",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "视图",
                        "name": "view",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "形状",
                        "name": "shape",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "边",
                        "name": "isEdge",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "是否常用",
                        "name": "common",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "调整大小",
                        "name": "resizing",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "分组",
                        "name": "group",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Must Target",
                        "name": "mustTarget",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "IDs",
                        "name": "ids",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/curd/cell/import": {
            "post": {
                "description": "导入",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cell管理"
                ],
                "summary": "导入",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer+空格+Token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "要上传的文件",
                        "name": "files",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/curd/cell/query": {
            "get": {
                "description": "分页查询",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Cell管理"
                ],
                "summary": "分页查询",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "从1开始",
                        "name": "pageIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "默认每页10条",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序字段",
                        "name": "orderField",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否倒序排序",
                        "name": "desc",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Must Source",
                        "name": "mustSource",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "系统",
                        "name": "system",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "视图",
                        "name": "view",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "形状",
                        "name": "shape",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "边",
                        "name": "isEdge",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "是否常用",
                        "name": "common",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "调整大小",
                        "name": "resizing",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "分组",
                        "name": "group",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Must Target",
                        "name": "mustTarget",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.QueryCellResponse"
                        }
                    }
                }
            }
        },
        "/api/curd/cell/update": {
            "put": {
                "description": "更新",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cell管理"
                ],
                "summary": "更新",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Update Cell",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/curd.CellInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "curd.CellInfo": {
            "type": "object",
            "properties": {
                "attrs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/curd.CellAttrs"
                    }
                },
                "common": {
                    "type": "boolean"
                },
                "connectings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/curd.CellConnecting"
                    }
                },
                "defaultEdge": {
                    "type": "boolean"
                },
                "defaultLabel": {
                    "type": "string"
                },
                "defaultLabelAttrs": {
                    "type": "string"
                },
                "formDefaultValue": {
                    "description": "表单默认值,json格式",
                    "type": "string"
                },
                "group": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "height2": {
                    "type": "integer"
                },
                "icon": {
                    "type": "string"
                },
                "iconSource": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "idPrefix": {
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                },
                "isEdge": {
                    "type": "boolean"
                },
                "isMust": {
                    "description": "系统必须要有的数据",
                    "type": "boolean"
                },
                "markup": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/curd.CellMarkup"
                    }
                },
                "mustSource": {
                    "type": "boolean"
                },
                "mustTarget": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "other": {
                    "description": "其他属性",
                    "type": "string"
                },
                "parent": {
                    "type": "boolean"
                },
                "ports": {
                    "type": "string"
                },
                "projectID": {
                    "type": "string"
                },
                "propertyForm": {
                    "type": "string"
                },
                "resizing": {
                    "type": "boolean"
                },
                "shape": {
                    "type": "string"
                },
                "system": {
                    "type": "string"
                },
                "tenantID": {
                    "type": "string"
                },
                "view": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                },
                "width2": {
                    "type": "integer"
                }
            }
        },
...
        "curd.CommonResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "curd.DelRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "tenantID": {
                    "type": "string"
                }
            }
        },
        "curd.EnableRequest": {
            "type": "object",
            "properties": {
                "enable": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "tenantID": {
                    "type": "string"
                }
            }
        },
...
        "curd.GetAllCellResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/curd.CellInfo"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "curd.GetCellDetailResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/curd.CellInfo"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}